backend_v2/requirements.txt
# Core FastAPI Application Dependencies
fastapi
uvicorn[standard]
python-dotenv
pydantic

# LangGraph and LangChain for Multi-Agent System
langgraph
langchain-openai
langchain-core
langchain-community

# Tavily API for Web Search
tavily-python

# Visualization Tools
plotly
kaleido

# Optional: Mind map generation (if needed)
pydot
graphviz
networkx

# Optional: Database (MongoDB)
motor
pymongo
dnspython
certifi

# AWS S3 Integration
boto3

# Production Server
gunicorn

# Testing
pytest
pytest-asyncio
httpx


# Data manipulation and Excel export
pandas
openpyxl

# Template Rendering (for SitRep HTML generation)
jinja2

# PDF Generation (for SitRep PDF reports)
pdfkit

# Social Media Content Tools (Infographic, Reel, Deck generators)
Pillow
moviepy
imageio-ffmpeg
python-pptx
